requirements/pipeline_requirements.txt
lightning==2.2.5
numpy==1.26.4
omegaconf==2.3.0
hydra-core==1.3.2
wandb==0.17.5
matplotlib==3.9.0
h5py==3.11.0
# for generating the images
opencv-python-headless
scipy
pyrender==0.1.45
#pyopengl==3.1.4 # NOTE: pip install this after the setup. This is a work around to avoid the error: cannot import name 'OSMesaCreateContextAttribs', ignore the compatibility error
